requirements.txt
fastapi
transformers
torch
Pillow